adds remote cert checking, and uses attachments
this upgrades CheckYourSSeLf to do two new things.

1. First CheckYourSSeLf can now check remote urls for expiring certs.
Because maybe not all of your certificates are in AWS, and that's okay.
Maybe you just wanna see if something crazy happens and google lets
their cert expire.

2. CheckYourSSeLf also now uses attachements. This makes the output
a lot more easier to read in the long run, and can quickly draw attetion
to the ones in danger with colors.